requirements.txt
Django>=4.2,<5
djangorestframework
django-cors-headers
celery
redis
beautifulsoup4
requests
selenium
scrapy
pandas
numpy
scikit-learn
transformers
torch
nltk
spacy
django-channels
channels-redis
django-allauth
django-crispy-forms
crispy-bootstrap5
django-import-export
django-filter
django-notifications-hq
channels
websockets
python-dotenv
psycopg2-binary
# Add this line to requirements.txt
webdriver-manager